eight commit

+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -16,6 +16,5 @@ package com.kodilla.testing.shape;
 
 public interface Shape {
     String getShapeName();
-    int getField();
-
+    double getField();
 }
